introduce a simple readiness probe that doesn't check bootstrapped status
Recently we have seen on mondaynet the node's RPC subsystem become
unresponsive, but the node does not crash.

We normally have a readiness probe to get alerted when this happens.

But the readiness probe is overkill: it checks whether the chain is
bootstrapped by measuring the age of the head block, failing if it's
over 10 minutes.

We don't want this on test networks generally, but especially on
mondaynet. After activation, we wait for the website to be published,
other participants to come online, and quorum to be met.

If we had this probe, the chain would be marked as unbootstrapped and
stop responding to RPC and p2p, then we would never get quorum.

But we still want to be alerted when the RPC subsystem is down.

I'm introducing 2 readiness probe settings:

* `bootstrapped_readiness_probe`: identical to existing
  `readiness_probe`
* `rpc_readiness_probe`: checks for RPC only

By default, they are on. So for mondaynet, the following should be set:

```
nodes:
  nodex:
    bootstrapped_readiness_probe: false
```